+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.8.0] - 2026-09-16
+
 ### Removed
 
 - **BREAKING: this package no longer provides an Account Center.** django-mvp